+++ src/ble/ctrlm_ble_rcu_interface.cpp
@@ -544,18 +544,10 @@ bool ctrlm_ble_rcu_interface_t::handleAddedDevice(const BleAddress &address)
 
             ctrlm_hal_ble_RcuStatusData_t params;
 
-            // Get supported vendor types from IR protocol bitmask value
-            int num_irdbs_supported = 0;
-            for (uint8_t bit = 0; bit < 8; bit++) {
-                if (irSupport & (0x1 << bit)) {
-                    params.rcu_data.irdbs_supported[num_irdbs_supported] = ctrlm_ble_utils_IrControlToVendor(0x1 << bit);
-                    num_irdbs_supported++;
-                }
-            }
+            params.rcu_data.irdbs_supported = irSupport;
 
             params.property_updated = CTRLM_HAL_BLE_PROPERTY_IRDBS_SUPPORTED;
             params.rcu_data.ieee_address = address.toUInt64();
-            params.rcu_data.num_irdbs_supported = num_irdbs_supported;
 
             m_rcuStatusChangedSlots.invoke(&params);
         };
@@ -678,15 +670,7 @@ ctrlm_hal_ble_rcu_data_t ctrlm_ble_rcu_interface_t::getAllDeviceProperties(uint6
     }
 
     // IR Service
-    uint8_t irSupport = device->irSupport();
-    int num_irdbs_supported = 0;
-    for (uint8_t bit = 0; bit < 8; bit++) {
-        if (irSupport & (0x1 << bit)) {
-            ret.irdbs_supported[num_irdbs_supported] = ctrlm_ble_utils_IrControlToVendor(0x1 << bit);
-            num_irdbs_supported++;
-        }
-    }
-    ret.num_irdbs_supported = num_irdbs_supported;
+    ret.irdbs_supported = device->irSupport();
 
     // Voice Service
     ret.audio_codecs = device->audioCodecs();
@@ -1138,7 +1122,7 @@ bool ctrlm_ble_rcu_interface_t::getFirstAudioDataTime(uint64_t ieee_address, ctr
     return false;
 }
 
-bool ctrlm_ble_rcu_interface_t::setIrControl(uint64_t ieee_address, ctrlm_irdb_vendor_t vendor)
+bool ctrlm_ble_rcu_interface_t::setIrControl(uint64_t ieee_address, uint8_t vendor)
 {
     // This method will wait for the operation to complete, so init a semaphore
     sem_t semaphore;
@@ -1163,10 +1147,9 @@ bool ctrlm_ble_rcu_interface_t::setIrControl(uint64_t ieee_address, ctrlm_irdb_v
 
     if (m_controller) {
         const auto device = m_controller->managedDevice(address);
-        uint8_t irControl = ctrlm_ble_utils_VendorToIrControlBitmask(vendor);
         if (device) {
             sem_init(&semaphore, 0, 0);
-            device->setIrControl(irControl, PendingReply<>(m_isAlive, replyHandler));
+            device->setIrControl(vendor, PendingReply<>(m_isAlive, replyHandler));
         } else {
             return false;
         }
@@ -1182,8 +1165,8 @@ bool ctrlm_ble_rcu_interface_t::setIrControl(uint64_t ieee_address, ctrlm_irdb_v
 }
 
 bool ctrlm_ble_rcu_interface_t::programIrSignalWaveforms(uint64_t ieee_address, 
-                                                         ctrlm_irdb_ir_codes_t &&irWaveforms, 
-                                                         ctrlm_irdb_vendor_t vendor)
+                                                         std::map<ctrlm_key_code_t, std::vector<uint8_t>> &&irWaveforms, 
+                                                         uint8_t vendor)
 {
     // lambda invoked when the request returns
     auto replyHandler = [this](PendingReply<> *reply) mutable
@@ -1210,7 +1193,6 @@ bool ctrlm_ble_rcu_interface_t::programIrSignalWaveforms(uint64_t ieee_address,
     XLOGD_INFO("programming IR waveforms on remote %s", address.toString().c_str());
 
     bool success = true;
-    uint8_t irControl = ctrlm_ble_utils_VendorToIrControlBitmask(vendor);
 
     if (m_controller) {
         const auto device = m_controller->managedDevice(address);
@@ -1258,7 +1240,7 @@ bool ctrlm_ble_rcu_interface_t::programIrSignalWaveforms(uint64_t ieee_address,
                 bleRcuWaveforms[key_code] = std::move(waveform.second);
             }
 
-            device->programIrSignalWaveforms(std::move(bleRcuWaveforms), irControl, 
+            device->programIrSignalWaveforms(std::move(bleRcuWaveforms), vendor, 
                     PendingReply<>(m_isAlive, replyHandler) );
             
         } else {