+++ src/ble/ctrlm_ble_utils.h
@@ -52,8 +52,6 @@ typedef struct {
 
 uint16_t ctrlm_ble_utils_ConvertUsbKbdCodeToLinux(unsigned char usb_code);
 unsigned char ctrlm_ble_utils_ConvertLinuxCodeToUsbKdb(uint16_t linux_code);
-uint8_t ctrlm_ble_utils_VendorToIrControlBitmask(ctrlm_irdb_vendor_t vendor);
-ctrlm_irdb_vendor_t ctrlm_ble_utils_IrControlToVendor(uint8_t irControl);
 
 std::string ctrlm_ble_utils_BuildDBusDeviceObjectPath(const char *path_base, unsigned long long ieee_address);
 unsigned long long ctrlm_ble_utils_GetIEEEAddressFromObjPath(std::string obj_path);
@@ -61,6 +59,5 @@ void ctrlm_ble_utils_PrintRCUStatus(ctrlm_hal_ble_RcuStatusData_t *status);
 const char *ctrlm_ble_unpair_reason_str(ctrlm_ble_RcuUnpairReason_t reason);
 const char *ctrlm_ble_reboot_reason_str(ctrlm_ble_RcuRebootReason_t reason);
 const char *ctrlm_ble_rcu_action_str(ctrlm_ble_RcuAction_t reason);
-std::string ctrlm_ble_irdbs_supported_str(std::vector<ctrlm_irdb_vendor_t> vendors);
 
 #endif